Add files via upload
Array Manager is a program in which given data, it assigns it to specific frames in memory which reside in the free frames list.  An array is then created with a corresponding page table.  This Array has the ability to be aliased, deleted, and resized.  The Array Client inputs characters corresponding to the desired action, "D" deletes the given array number, "C" creates an array of a given size, "S" sets a given array, and so forth.